Cleaned up passing of history/handleSearch function on each Pagination rendering
- Also some esling fixes
