+++ src/Utilities/helpers.js
@@ -1,11 +1,4 @@
 import moment from 'moment';
-import {
-  encodeNonDefaultQueryString,
-  mergeParams,
-  parseQueryString,
-  removeParams,
-  replaceParams,
-} from './qs';
 
 /*
  * isNumeric - return true if input is a real number
@@ -144,32 +137,3 @@ export function formatJobDetailsURL(baseURL, jobId) {
   const subDirectory2 = 'details';
   return `${baseURL}/#/${subDirectory1}/${jobId}/${subDirectory2}/`;
 }
-
-// export const handleSearch = (key, value, qsConfig, history) => {
-//   let params = parseQueryString(qsConfig, history.location.search);
-//   params = mergeParams(params, { [key]: value });
-//   //params = replaceParams(params, { page: 1 });
-//   if (value === '' || value.length === 0)
-//     params = removeParams(qsConfig, params, {[key]: params[key]});
-//   pushHistoryState(params, qsConfig);
-// }
-
-export const handleSearch = (key, value, qsConfig, history) => {
-  let params = parseQueryString(qsConfig, history.location.search);
-  params = replaceParams(params, { [key]: value });
-  params = mergeParams(params, { [key]: value });
-  if (value === '' || value.length === 0)
-    params = removeParams(qsConfig, params, { [key]: params[key] });
-  pushHistoryState(params, qsConfig, history);
-};
-
-export const pushHistoryState = (params, qsConfig, history) => {
-  const { pathname } = history.location;
-  const nonNamespacedParams = parseQueryString({}, history.location.search);
-  const encodedParams = encodeNonDefaultQueryString(
-    qsConfig,
-    params,
-    nonNamespacedParams
-  );
-  history.push(encodedParams ? `${pathname}?${encodedParams}` : pathname);
-};
